Fix build warnings for wifi_cert, wifi_cli, wpa_supplicant
[Description]
warning: equality comparison with extraneous parentheses [-Wparentheses-equality]
note: remove extraneous parentheses around the comparison to silence this warning

[Root Cause]
The compiler's Pe546 warning is intended to prevent bypassing the initialization
of local variables in jump statements (such as goto), which may cause these
variables to be undefined when subsequently used.

[Fix]
Declare and initialize variables in advance, Move the variable declaration
and initialization up to before the goto statement.

[Testing]
compile ok

[Reviewers]
Gannicus Guo, Supriya Sumeet Inamdar, Qingling Wu

[Port proposal]
non-general

+++ wifidriver/mlan_api.c
@@ -6476,7 +6476,7 @@ static void proc_csi_event(void *p_data)
             }
             mlan_adap->ami_num++;
 			PRINTF("NUM %d CSI Processing Results: %s(%d), RX/TX %d/%d, %-8.2f TSF %llx, Ambient Motion Index %0.1f dB\r\n",
-				mlan_adap->ami_num, myStr, l_BW, nRx, nTx, toa_ns, TSF, ambientMotionVal_dB);
+				mlan_adap->ami_num, myStr, l_BW, nRx, nTx, (double)toa_ns, TSF, (double)ambientMotionVal_dB);
             
             if (g_ami_cfg.gcsi_filter_param.num_csi)
 			{
